pkg/apis/kf/v1alpha1: overhaul routing controllers
Previous to this CL, a Route object knew about which every App
(`AppNames []string`) it had to route traffic to. This implied that it
also had to watch for Apps being deleted. The App also knew about the
RouteSpecFields, and therefore the App controller had to merge Routes
when necessary. This could be complicated and had several edge cases.

This CL changes it to where each Route only knows about one App (or none
in the case of a special route called a 'claim'). For every route that
an App wants (via its `App.Spec.Routes`), a new Route is created. For
every new Route (`Hostname`+`Domain`+`Path` combo), a new Route claim is
created. Therefore the route controller can aggregate the routes based
on their labels and create the necessary Istio VirtualServices.

Route Controls:
- create route: Create Route claim only.
- map route: Append to `App.Spec.Routes`.
- unmap route: Delete from `App.Spec.Routes`.
- Delete app: App owns the Route, so its automatically cleaned up.
- Delete Route: Unmap each App that has the Route, then delete Route
claim.

fixes #417